+++ private/test/src/matrix-node/embedded/v2/instrumentation.test.ts
@@ -18,7 +18,7 @@ testInstrumentation({
       policies: readFile(
         resolve(__dirname, "../../../../bundles/PS2MX9855QURB3Y8.crrt"),
       ),
-      wasm: readFile(require.resolve("@cerbos/embedded-server/wasm")),
+      wasm: readFile(require.resolve("@cerbos/embedded-server/server.wasm")),
       decodeJWTPayload: ({ token }): DecodedJWTPayload =>
         UnsecuredJWT.decode(token).payload as DecodedJWTPayload,
       globals: {